+++ image.go
@@ -1,11 +1,7 @@
 package rui
 
 import (
-	"encoding/base64"
-	"path/filepath"
-	"runtime"
 	"strconv"
-	"strings"
 )
 
 const (
@@ -83,27 +79,6 @@ func (manager *imageManager) loadImage(url string, onLoaded func(Image), session
 	image.loadingStatus = ImageLoading
 	manager.images[url] = image
 
-	if runtime.GOOS == "js" && wasmMediaResources {
-		if file, ok := resources.images[url]; ok && file.fs != nil {
-
-			dataType := map[string]string{
-				".svg":  "data:image/svg+xml",
-				".png":  "data:image/png",
-				".jpg":  "data:image/jpg",
-				".jpeg": "data:image/jpg",
-				".gif":  "data:image/gif",
-			}
-			ext := strings.ToLower(filepath.Ext(url))
-			if prefix, ok := dataType[ext]; ok {
-				if data, err := file.fs.ReadFile(file.path); err == nil {
-					session.callFunc("loadInlineImage", url, prefix+";base64,"+base64.StdEncoding.EncodeToString(data))
-					return image
-				} else {
-					DebugLog(err.Error())
-				}
-			}
-		}
-	}
 	session.callFunc("loadImage", url)
 	return image
 }